Task 1 Level 1 Landing Page
Welcome to the CodSoft repository! This repository contains the code for my web development internship projects. The projects are organized into three main tasks, each stored in its Separate File.

File Structure :

Name Are Written In front of file of which it was.

Task 1: Landing Page Design This folder contains the code for a landing page. The design showcases my ability to create appealing and responsive landing pages using HTML and CSS.

Task 2: Portfolio Design This folder contains the code for a personal portfolio website. The design is implemented using HTML and CSS.

Task 3: Calculator Design This folder contains the code for a basic calculator application. The design and functionality are implemented using HTML, CSS, and JavaScript.